Refactor form range functionality for reuse on book search form
