+++ mep/common/forms.py
@@ -209,3 +209,34 @@ class FacetForm(forms.Form):
                     for val, count in counts.items()
                 ]
                 self.fields[formfield].widget.facet_counts = counts
+
+
+class RangeForm(forms.Form):
+    '''Form mixin to initialize min/max values for range fields.'''
+
+    def set_range_minmax(self, range_minmax):
+        '''Set the min, max, and placeholder values for all
+        :class:`~mep.common.forms.RangeField` instances.
+
+        :param range_minmax: a dictionary with form fields as key names and
+            tuples of min and max integers as values.
+        :type range_minmax: dict
+
+        :rtype: None
+        '''
+        for field_name, min_max in range_minmax.items():
+            self.fields[field_name].set_min_max(min_max[0], min_max[1])
+
+    def __init__(self, data=None, *args, **kwargs):
+        '''
+        Override to set choices dynamically and configure min-max range values
+        based on form kwargs.
+        '''
+        # pop range_minmax out of kwargs to avoid clashing
+        # with django args
+        range_minmax = kwargs.pop('range_minmax', {})
+
+        super().__init__(data=data, *args, **kwargs)
+
+        # call function to set min_max and placeholders
+        self.set_range_minmax(range_minmax)
